+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Thank you all
 Antonio Ramirez.
 
 ### YiiBooster version 1.0.7
+- **(enh)** Added the possibility to render append/prepend without a span wrap #414 (Frostmaind)
 - **(enh)** Update x-editable-yii from 1.3.0 to 1.4.0 #413 (magefad)